// src/types/enums/userProfileEnum.tsx
export enum EGenderSelection {
    dr = 'Dr.',
    mr = 'Mr.',
    ms = 'Ms.',
    mrs = 'Mrs.',
    mx = 'Mx.',
    none = 'None',
}

export enum EGenderNames {
    male = 'Male',
    female = 'Female',
    unspecified = 'Unspecified',
    none = 'none'
}

export enum EProfileColours {
    green = 'Green',
    blue = 'Blue',
    red = 'Red',
    defualt = 'Grey'
}